Removed pointer to next state to reduce ram load.
Since the goal of the project is to optimize the performance of the script and not to create a usable tool there's no need to save the gamestate graph, so the pointers are not relevant bcs. the next gamestates are just put onto the queue while the current one gets discarded.